+++ app/settings/contentCard-settings-04-features.php
@@ -1,0 +1,18 @@
+<?php
+// Card
+$cardHeaderContent = \xan\fontIcon( 'fas fa-sliders-h-square' ) . STR_NBSP . 'Features';
+$card = new \xan\eleCard( CARD_WIDTH, '', false );
+
+// Tags Special
+
+// Table
+$table = new \xan\eleTable( $tagsCellEmpty );
+$tableRowIndex = -1;
+
+// Table Rows
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmSettingsT->getColEleRender( 'LogoutAutoSeconds', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmSettingsT->getColEleRender( 'LogoutAutoSeconds', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+// Card Append
+$resp->contentAreaA[] = $card->renderCardWithDiv( $cardHeaderContent, $table->render() );
+?>